pyproject.toml
[tool.poetry]
name = "advanced_rag"
version = "0.1.0"
description = "Advanced RAG System using Ollama"
authors = ["Geetheswar <142201025@smail.iitpkd.ac.in>",
            "Bhavani Shankar <112201026@smail.iitpkd.ac.in>",
            "Shiva Shankar <142201028@smail.iitpkd.ac.in>"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
qdrant-client = "^1.7.0"
beautifulsoup4 = "^4.12.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
ragas = "^0.1.0"
evaluate = "^0.4.1"
nltk = "^3.8.1"
pandas = "^2.1.3"
aiohttp = "^3.9.1"
tenacity = "^8.2.3"
pydantic = "^2.5.2"
tqdm = "^4.66.1"
jinja2 = "^3.1.2"
lxml = "^4.9.3"
transformers = "^4.35.2"
arxiv = "^2.0.0"
typer = "^0.9.0"
rich = "^13.7.0"
datasets = "^2.15.0"
absl-py = "^2.2.2"
rouge-score = "^0.1.2"
python-docx = "^1.1.2"
python-multipart = "^0.0.20"
pymupdf = "^1.25.5"
sentence-transformers = "^4.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
